pyproject.toml
[tool.pytest.ini_options]
pythonpath = [".", "src",]
asyncio_mode = "auto"

[tool.poetry]
name = "first-proj"
version = "0.1.0"
description = ""
authors = ["awesomex <bestawes0mex@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
gunicorn = "^22.0.0"
uvicorn = "^0.30.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
python-multipart = "^0.0.9" 

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
